requirements.txt
# Core dependencies
python-dotenv==1.0.1
requests==2.31.0

# OpenAI
openai==1.12.0

# WordPress
python-wordpress-xmlrpc==2.3
Markdown==3.5.2

# Google Sheets
gspread==6.2.0
oauth2client==4.1.3

# Image processing
Pillow==10.2.0

# Utility packages
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==8.3.5
pytest-mock==3.12.0

# Development
black==24.2.0
flake8==7.0.0
isort==5.13.2 